+++ src/component/AddPage.js
@@ -1,0 +1,5 @@
+function AddPage(){
+    alert("click page!");
+}
+
+export default AddPage;
